micro_services/silero_vad/pyproject.toml
[project]
name = "vixio-silero-vad-service"
version = "1.0.0"
description = "Silero VAD gRPC microservice for Vixio"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "silero-vad>=5.0",
    "loguru>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

# UV package indexes (optional, for faster downloads in China)
# [[tool.uv.index]]
# url = "https://mirrors.aliyun.com/pypi/simple/"
# default = true

